backend/requirements.txt
Django==3.2.3
djoser==2.1.0
djangorestframework==3.12.4
django-filter==22.1
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.3
Pillow==9.0.0
flake8-isort==6.0.0
flake8==5.0.4